+++ src/app/map/map.component.ts
@@ -1,8 +1,7 @@
 import { CommonModule } from '@angular/common';
-import { Component, ElementRef, Input, OnDestroy, OnInit, effect, input, viewChild, inject } from '@angular/core';
-import { PinchZoomModule } from '@meddv/ngx-pinch-zoom';
+import { Component, ElementRef, Input, OnDestroy, OnInit, effect, input, viewChild, inject, output } from '@angular/core';
 import { LocationEnabledStatus, MapInfo, MapPoint, Pin } from '../data/models';
-import { defaultMapRadius, distance, formatDistanceMiles, mapPointToPin } from './map.utils';
+import { calculateRelativePosition, defaultMapRadius, distance, formatDistanceNice, mapPointToPin } from './map.utils';
 import { delay } from '../utils/utils';
 import { GeoService } from '../geolocation/geo.service';
 import { SettingsService } from '../data/settings.service';
@@ -17,25 +16,14 @@ import { DbService } from '../data/db.service';
 import { MapModel, MapResult } from './map-model';
 import { init3D } from './map';
 
-interface MapInformation {
-  width: number; // Width of the map
-  height: number; // Height of the map
-  circleRadius: number; // Half width
-}
-
 // How often is the map updated with a new location
 const geolocateInterval = 10000;
-
-// Offset x,y in pixel of the "you are here" pin
-const youOffsetX = 6;
-const youOffsetY = 4;
 
 @Component({
   selector: 'app-map',
   templateUrl: './map.component.html',
   styleUrls: ['./map.component.scss'],
   imports: [
-    PinchZoomModule,
     RouterModule,
     CommonModule,
     MessageComponent,
@@ -62,6 +50,7 @@ export class MapComponent implements OnInit, OnDestroy {
   showMessage = false;
   hideCompass = false;
   pointsSet = false;
+  selectedPoint: MapPoint | undefined;
   pins: Pin[] = [];
   private geoInterval: any;
   private nearestPoint: number | undefined;
@@ -75,6 +64,8 @@ export class MapComponent implements OnInit, OnDestroy {
   footerPadding = input<number>(0);
   smallPins = input<boolean>(false);
   isHeader = input<boolean>(false);
+  scrolled = output<number>();
+
 
   @Input() set points(points: MapPoint[]) {
     if (this.pointsSet) return;
@@ -89,10 +80,20 @@ export class MapComponent implements OnInit, OnDestroy {
   }
 
   private async fixGPSAndUpdate() {
+    this.selectedPoint = undefined;
+    let foundPoints = 0;
     for (let point of this._points) {
       if (!point.gps) {
         point.gps = await this.geo.getMapPointToGPS(point);
       }
+      if (point.animated) {
+        this.selectedPoint = point;
+        foundPoints++;
+      }
+    }
+    if (foundPoints > 1) {
+      // Only select one point
+      this.selectedPoint = undefined;
     }
     this.hideCompass = !await this.db.hasGeoPoints();
     await delay(150);
@@ -177,7 +178,9 @@ export class MapComponent implements OnInit, OnDestroy {
     if (degree < 0) {
       degree += 360;
     }
-    console.log(`Compass heading is ${degree}`);
+
+    // Because the label shows left/right ahead we need to update the location
+    this.displayYourLocation(this.geo.gpsPosition());
     if (this.mapResult) {
       this.mapResult.rotateCompass(degree);
     }
@@ -193,11 +196,10 @@ export class MapComponent implements OnInit, OnDestroy {
     }
 
     if (this.mapResult) {
-      console.log('set my position', pt, gpsCoord);
       this.mapResult.myPosition(pt.x, pt.y);
     }
 
-    await this.calculateNearest(gpsCoord);
+    await this.calculateNearest(gpsCoord, this.geo.heading());
   }
 
   private setMapInformation() {
@@ -221,7 +223,10 @@ export class MapComponent implements OnInit, OnDestroy {
       pinClicked: this.pinClicked.bind(this),
     }
 
-    const blink = this.points.length == 1;
+    if (this.points.length == 1) {
+      this.points[0].animated = true;
+    }
+
     for (const [i, point] of this._points.entries()) {
       const pin = mapPointToPin(point, defaultMapRadius);
 
@@ -229,14 +234,20 @@ export class MapComponent implements OnInit, OnDestroy {
         map.pins.push({
           uuid: `${i}`,
           x: pin.x, z: pin.y,
-          color: point.info?.bgColor ?? 'primary', animated: blink, size: map.defaultPinSize, label: point.info?.label ?? '^'
+          color: point.info?.bgColor ?? 'primary',
+          animated: point.animated,
+          size: map.defaultPinSize,
+          label: point.info?.label ?? '^'
         });
       } else {
         console.error(`Point could not be converted to pin`);
       }
     }
-    console.log('init3D', map)
+
     this.mapResult = await init3D(this.container().nativeElement, map);
+    this.mapResult.scrolled = (deltaY: number) => {
+      this.scrolled.emit(deltaY);
+    }
     this._viewReady = true;
     await this.checkGeolocation();
     this.setupCompass();
@@ -298,12 +309,13 @@ export class MapComponent implements OnInit, OnDestroy {
     }
   }
 
-  private async calculateNearest(you: GpsCoord) {
+  private async calculateNearest(you: GpsCoord, compass: CompassHeading) {
     let least = 999999;
     let closest: MapPoint | undefined;
     let closestIdx = 0;
+    let name = 'Closest point';
     let idx = 0;
-    for (let point of this._points) {
+    for (let point of this.selectedPoint ? [this.selectedPoint] : this._points) {
       if (!point.gps || !point.gps.lat) {
         if (!environment.production) {
           console.error(`MapPoint is missing gps coordinate: ${JSON.stringify(point)}`);
@@ -315,6 +327,7 @@ export class MapComponent implements OnInit, OnDestroy {
           least = dist;
           closest = point;
           closestIdx = idx;
+          name = point.info?.title.replace('My ', 'Your ') ?? '';
         }
       }
       idx++;
@@ -332,17 +345,18 @@ export class MapComponent implements OnInit, OnDestroy {
         this.footer = ``;
         return;
       }
-      const prefix = this._points.length > 1 ? 'Closest point is ' : '';
-      const dist = formatDistanceMiles(least);
+      const prefix = this._points.length > 1 ? `${name} is ` : '';
+      const dist = formatDistanceNice(least);
       if (least > 50) {
         if (this.settings.settings.locationEnabled === LocationEnabledStatus.Enabled) {
-          this.footer = 'You are outside of the Event';
+          this.footer = (you.lat == 0) ? 'Calculating location' : 'You are outside of the Event';
         } else {
           this.footer = this.disabledMessage;
         }
       } else {
         if (dist != '') {
-          this.footer = `${prefix}${dist} away.`;
+          const direction = closest.gps ? calculateRelativePosition(you, closest.gps, compass.trueHeading) : 'away';
+          this.footer = `${prefix}${dist} ${direction}`;
         } else {
           this.footer = ``;
           console.error(`Unable to find a close point`);